exam scoring functionality is checked

+++ StayHard/Models/ExamParticipant.cs
@@ -1,10 +1,12 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace StayHard.Models;
 
 public class ExamParticipant
 {
     public int ExamId { get; set; }
     public int StudentId { get; set; }
-    public int? Score { get; set; }
+    public float? Score { get; set; }
     public Student Student { get; set; }
     public Exam Exam { get; set; }
 }
